requirements.txt
pandas
matplotlib
numpy
Pillow
tqdm
requests